requirements.txt
streamlit
google-generativeai
plotly==5.18.0
PyPDF2